feat(model): apply the instance of the relationship in the nested objects
This commit extract the code of `find` and `get` that applies the response data in the current instance to two methods `_applyInstance` and `_applyInstanceCollection`. It also add two new methods: `relations` that contains an object of the relations and `_applyRelations` that is used in `_applyInstance` to apply the instance of the relationship in the nested objects, based on `relations`.

Based on @Dylan-Chapman code.