feat(model): add support to configuration at query builder (#142)
Add a new method `config` to `Model.js`. It can be used to override the http instance (e.g. axios) configuration at query builder.
It can be used to change the method of update request to `PATCH`, to add additional data to the request, to add headers, and so on.